Pull out the PIC, DMA.
